Add a method for installation via apt
Apt is inherently more secure than downloading an archive on a website (without
verification). The packages are signed by the gravitational key, and we already
know their public key.  The compromise ot the key is less probable than the
infrastructure of gravitational fails.

Update is simpler as it does not needs a whole playbook.

Alternative methods may be implemented for others packages managers (yum,
zipper, dnf at this time).